+++ views/homepage.handlebars
@@ -62,6 +62,7 @@
 <script src ="javascript/create-account.js"></script>
 <script src ="javascript/create-owner.js"></script>
 <script src="/javascript/login.js"></script>
+<script src="/javascript/login-walker.js"></script>
 
 
 {{!-- {{> modal/login-block}} --}}
