+++ views/partials/jobs-card.handlebars
@@ -8,12 +8,12 @@
         <h5 class="card-text">Pay: {{ this.pay }}</h5>
         <h6 class="card-text text-muted">Time: {{ this.timeframe }}</h6>
         <h6 class="card-text text-muted">Location: {{ this.location }}</h6>
-        {{!-- (HELPER) function that determines if the person logged in is the creator of this job and it's active--}}
+        {{!-- {{#if owner}}
         {{#unless completed}}
         <button id="delete-job" class="float-right btn success" type="button">Delete</button>
         {{/unless}}
-        {{!-- (HELPER) if logged in as walker and job not accepted --}}
-        {{#if acceptable}}
+        {{/if}} --}}
+        {{#if walker}}
         <button id="accept-job" class="float-right btn success" type="button">Accept</button>
         {{/if}}
     </div>
