+++ views/partials/nav-block.handlebars
@@ -29,4 +29,6 @@
 
 <script src="/javascript/login.js"></script>
 <script src="/javascript/logout.js"></script>
+<script src="/javascript/login-walker.js"></script>
+
